[Refactor] 역할 분리를 위해 AppleOAuthUserProvider 클래스 생성
